requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0

sentence-transformers==2.7.0
transformers==4.41.0
torch==2.2.2

faiss-cpu==1.7.4
numpy==1.26.4

pydantic==2.7.1
scikit-learn==1.5.0

python-docx==1.1.2
PyMuPDF==1.24.1
